requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
certifi=2024.12.14=py310h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.2=h5eee18b_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py310h06a4308_1
jpeg=9e=h5eee18b_3
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.30=0
libnvfatbin=12.6.77=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py310h5eee18b_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py310h06a4308_0
numpy=2.0.1=py310h5f9d8c6_1
numpy-base=2.0.1=py310hb5e798b_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
pillow=11.0.0=py310hcea889d_1
pip=24.2=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.16=he870216_1
pytorch=2.5.1=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py310h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py310h06a4308_1
scipy=1.14.1=py310h5f9d8c6_0
setuptools=75.1.0=py310h06a4308_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
torchaudio=2.5.1=py310_cu124
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu124
tqdm=4.66.5=py310h2f386ee_0
typing_extensions=4.12.2=py310h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py310h06a4308_0
wheel=0.44.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
